+++ packages/graph/attachments/types.ts
@@ -1,6 +1,6 @@
 import { Attachment as IAttachmentType } from "@microsoft/microsoft-graph-types";
 import { body } from "@pnp/queryable";
-import { _GraphQueryableCollection, _GraphQueryableInstance, graphInvokableFactory } from "../graphqueryable.js";
+import { _GraphCollection, _GraphInstance, graphInvokableFactory } from "../graphqueryable.js";
 import { graphPost } from "../operations.js";
 import { defaultPath, getById, IGetById } from "../decorators.js";
 import { type } from "../utils/type.js";
@@ -8,7 +8,7 @@ import { type } from "../utils/type.js";
 /**
  * Attachment
  */
-export class _Attachment extends _GraphQueryableInstance<IAttachmentType> { }
+export class _Attachment extends _GraphInstance<IAttachmentType> { }
 export interface IAttachment extends _Attachment { }
 export const Attachment = graphInvokableFactory<IAttachment>(_Attachment);
 
@@ -17,7 +17,7 @@ export const Attachment = graphInvokableFactory<IAttachment>(_Attachment);
  */
 @defaultPath("attachments")
 @getById(Attachment)
-export class _Attachments extends _GraphQueryableCollection<IAttachmentType[]> {
+export class _Attachments extends _GraphCollection<IAttachmentType[]> {
 
     /**
      * Add attachment to this collection