+++ packages/graph/calendars/users.ts
@@ -2,7 +2,7 @@ import { addProp } from "@pnp/queryable";
 import { _User } from "../users/types.js";
 import { Calendar, ICalendar, IEvents, Events, Calendars, ICalendars } from "./types.js";
 import { calendarView, ICalendarViewInfo } from "./funcs.js";
-import { IGraphQueryableCollection } from "../graphqueryable.js";
+import { IGraphCollection } from "../graphqueryable.js";
 
 declare module "../users/types" {
     interface _User {
@@ -10,14 +10,14 @@ declare module "../users/types" {
         readonly calendars: ICalendars;
         readonly attachmentFiles: ICalendar;
         readonly events: IEvents;
-        calendarView(start: string, end: string): IGraphQueryableCollection<ICalendarViewInfo[]>;
+        calendarView(start: string, end: string): IGraphCollection<ICalendarViewInfo[]>;
     }
     interface IUser {
         readonly calendar: ICalendar;
         readonly calendars: ICalendars;
         readonly attachmentFiles: ICalendar;
         readonly events: IEvents;
-        calendarView(start: string, end: string): IGraphQueryableCollection<ICalendarViewInfo[]>;
+        calendarView(start: string, end: string): IGraphCollection<ICalendarViewInfo[]>;
     }
 }