+++ packages/graph/lists/types.ts
@@ -1,5 +1,5 @@
 import { List as IListEntity } from "@microsoft/microsoft-graph-types";
-import { _GraphQueryableCollection, graphInvokableFactory, _GraphQueryableInstance } from "../graphqueryable.js";
+import { _GraphCollection, graphInvokableFactory, _GraphInstance } from "../graphqueryable.js";
 import { defaultPath, deleteable, IDeleteable, updateable, IUpdateable, getById, IGetById } from "../decorators.js";
 import { graphPost } from "../operations.js";
 import { body } from "@pnp/queryable";
@@ -9,7 +9,7 @@ import { body } from "@pnp/queryable";
  */
 @deleteable()
 @updateable()
-export class _List extends _GraphQueryableInstance<IListEntity> { }
+export class _List extends _GraphInstance<IListEntity> { }
 export interface IList extends _List, IDeleteable, IUpdateable { }
 export const List = graphInvokableFactory<IList>(_List);
 
@@ -19,7 +19,7 @@ export const List = graphInvokableFactory<IList>(_List);
  */
 @defaultPath("services")
 @getById(List)
-export class _Lists extends _GraphQueryableCollection<IListEntity[]>{
+export class _Lists extends _GraphCollection<IListEntity[]>{
     /**
      * Create a new booking service as specified in the request body.
      *